mv grpc feature gated code under service
